requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
ultralytics>=8.0.0
timm>=0.9.0
nuscenes-devkit>=1.1.11
PyYAML>=6.0
scipy>=1.11.0
Pillow>=10.0.0
tqdm>=4.65.0
